Rename RunModel package and subpackage to be pep8 compliant
