+++ ApertureMapModelTools/run_model/bulk_run.py
